HistoryService/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
flask-restx==1.3.0
future==0.18.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
PyJWT==2.8.0
pytz==2024.1
referencing==0.33.0
rpds-py==0.18.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
